+++ src/main/kotlin/com/grandemc/fazendas/storage/player/model/FarmQuest.kt
@@ -1,12 +1,12 @@
 package com.grandemc.fazendas.storage.player.model
 
 class FarmQuest(
-    private val id: Byte,
+    private val id: Short,
     private val type: QuestType,
     private var progress: Int = 0,
     private var done: Boolean = false
 ) {
-    fun id(): Byte = id
+    fun id(): Short = id
     fun type(): QuestType = type
     fun progress(): Int = progress
     fun advance(points: Int) {
